# olympus/vehicle-api/openapi.yaml
openapi: 3.1.0
info:
  title: OLYMPUS OS Vehicle API
  description: >
    REST and WebSocket gateway for the OLYMPUS OS drone fleet coordination
    system. Provides vehicle management, mission control, detection tracking,
    telemetry metrics, webhook event delivery, and an AI advisor chat proxy.
    All protected endpoints require a Bearer token in the Authorization header.
  version: 1.0.0
  contact:
    name: OLYMPUS OS Team
servers:
  - url: http://localhost:3001
    description: Local development server
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/webhooks:
    post:
      operationId: registerWebhook
      summary: Register a webhook
      description: >
        Registers a new webhook endpoint for event delivery. The URL must
        use HTTPS. Allowed events: detection, task_assigned,
        task_completed, mission_start, mission_abort, vehicle_online,
        vehicle_offline. An optional secret is hashed (SHA-256) and used
        to sign outgoing payloads via the X-Olympus-Signature header.
      tags:
        - webhooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookRegistration"
      responses:
        "201":
          description: Webhook registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookRegisterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    get:
      operationId: listWebhooks
      summary: List webhooks
      description: >
        Returns webhooks belonging to the authenticated partner. Partners
        with the `*` scope can see all webhooks.
      tags:
        - webhooks
      responses:
        "200":
          description: Array of webhook entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WebhookEntry"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Advisor ─────────────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        API key passed as a Bearer token in the Authorization header.
        When no API keys are configured the server runs in dev mode with
        authentication disabled.

  # ── Reusable Parameters ─────────────────────────────────────────────
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - status
      properties:
        error:
          type: string
          description: Human-readable error message
        status:
          type: integer
          description: HTTP status code
          examples:
            - 400
            - 404
            - 409
            - 500
            - 502
...
    WebhookRegistration:
      type: object
      required:
        - url
        - events
      properties:
        url:
          type: string
          format: uri
          description: HTTPS callback URL
          pattern: "^https://"
        events:
          type: array
          items:
            type: string
            enum:
              - detection
              - task_assigned
              - task_completed
              - mission_start
              - mission_abort
              - vehicle_online
              - vehicle_offline
          minItems: 1
          description: Events to subscribe to
        secret:
          type: string
          nullable: true
          description: >
            Optional shared secret. When set, outgoing payloads are signed
            with HMAC-SHA256 and the signature is sent in the
            X-Olympus-Signature header.
    WebhookRegisterResponse:
      type: object
      required:
        - id
        - status
      properties:
        id:
          type: string
          format: uuid
          description: Unique webhook registration ID
        status:
          type: string
          description: Registration status
          const: registered
    WebhookEntry:
      type: object
      required:
        - id
        - partner_id
        - url
        - events
        - created_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique webhook ID
        partner_id:
          type: string
          description: Partner who registered this webhook
        url:
          type: string
          format: uri
          description: HTTPS callback URL
        events:
          type: array
          items:
            type: string
          description: Subscribed event types
        secret_hash:
          type: string
          nullable: true
          description: SHA-256 hash of the secret (never the raw secret)
        created_at:
          type: string
          format: date-time
          description: When the webhook was registered (RFC 3339)

    # -- Advisor --
...